requirements.txt
streamlit
python-dotenv
pymongo
serpapi
google-search-results
langchain-google-genai
langgraph
langchain-core